requirements.txt
apache-airflow==2.9.2
pycaret
scikit-learn
pandas
# joblib
joblib==1.2.0
kaggle


# apache-airflow==2.9.2
# pycaret==3.0.0
# scikit-learn
# pandas
# joblib
# kaggle
